feat(block reply) 1. Add nonce_ext support for height after enable Ethash PoW

+++ abejson/chainsvrresults.go
@@ -126,7 +126,7 @@ type GetBlockAbeVerboseResult struct {
 	Tx           []string         `json:"tx,omitempty"`
 	RawTx        []TxRawResultAbe `json:"rawtx,omitempty"` // Note: this field is always empty when verbose != 2.
 	Time         int64            `json:"time"`
-	Nonce        uint32           `json:"nonce"`
+	Nonce        uint64           `json:"nonce"`
 	Bits         string           `json:"bits"`
 	Difficulty   float64          `json:"difficulty"`
 	PreviousHash string           `json:"previousblockhash"`
